+++ src/com/reason/ide/spellcheckers/RmlSpellCheckerStrategy.java
@@ -4,9 +4,9 @@ import com.intellij.psi.*;
 import com.intellij.spellchecker.tokenizer.*;
 import org.jetbrains.annotations.*;
 
-public class ReasonSpellCheckerStrategy extends CustomSpellCheckerStrategy {
-
-    @Override public @NotNull Tokenizer<?> getTokenizer(PsiElement element) {
+public class RmlSpellCheckerStrategy extends ORSpellCheckerStrategy {
+    @Override
+    public @NotNull Tokenizer<?> getTokenizer(@Nullable PsiElement element) {
         return super.getTokenizer(element);
     }
 }